chg: Renamed to OSS
This is only the library for integrating into an application so the name
was changed to "sso" as it seemed more correct and specific to its
purpose.

Login will be the name of the "kohirens/www/login" package that provides
that implements the "sso" library functionality and provides
developers access to code to get started quickly, or an example of
possibilities.