chg: Data Stored in the Login Information
Added the account ID of the client that is logged in to the stored
login information. This is important because it allows retrieval of
a clients account given the provider client ID which can be recovered
from reconsent. This plays a major role in account de-duplication and
ensuring that the correct account is always used for authentication
and authorization.